factor out logsec into a class
